cgame: ffa1 trace clean — macro multi-eval rand draws + sin-chain f32 consts (wf #5)
The wide ffa1 trace exposed two fidelity classes the first three traces never exercised.

Raven's VectorMA/VectorScale are macros, so a rand-consuming scale argument draws once per component. Our port hoisted the draw at three sites: the CG_ForceElectrocution reach (cg_players.c:7850, proven bit-exact by an LCG stream search - our end vector sits at seed-1 draw 4060, the recorded one needs the per-component draws), and the two CG_Chunks VectorScale calls (cg_effects.c:1218,1227). The game module already ports this class correctly (NPC_AI_Seeker.rs:334, NPC_AI_Sniper.rs:709).

Four sin-flicker chains widened bare f64 literals where Raven writes f32 constants - (double)0.08f differs from 0.08, a one-ulp wobble the sprite alpha compare catches (cg_players.rs saber glow, cg_draw.rs zoom mask + corona).

The replay referee gains a dump_trace_window inspector test (JKA_DUMP=<lo>:<hi>) and full-precision float previews.

All four traces now replay both gates at 0 findings: ffa1 (5,479,080 records), swoop1 (33,759,450), sabers1, spectator.

+++ crates/mp/cgame/src/cg_draw.rs
@@ -3172,7 +3172,7 @@ pub fn CG_DrawZoomMask(ctx: &mut CgContext) {
             color1[0] = 1.0;
             color1[1] = 1.0;
             color1[2] = 1.0;
-            color1[3] = (0.7 + ((ctx.world.cg.time as f32 * 0.01) as f64).sin() * 0.3) as f32;
+            color1[3] = (0.7f32 as f64 + ((ctx.world.cg.time as f32 * 0.01) as f64).sin() * 0.3f32 as f64) as f32;
 
             trap::R_SetColor(ctx.engine, Some(&color1));
         }
@@ -6341,7 +6341,7 @@ pub fn CG_DrawCrosshair(ctx: &mut CgContext, worldPoint: Option<vec3_t>, chEntVa
         ecolor[3] = 0.5;
         // don't draw full color
         let v = ((1.0 - ecolor[3]) as f64
-            * (((ctx.world.cg.time as f32 * 0.001) as f64).sin() * 0.08 + 0.35))
+            * (((ctx.world.cg.time as f32 * 0.001) as f64).sin() * 0.08f32 as f64 + 0.35f32 as f64))
             as f32;
         ecolor[0] = v;
         ecolor[1] = v;
